pyproject.toml
[project]
name = "paperbanana"
version = "0.1.0"
description = "Automated Academic Illustration Generation Framework"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.1.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "python-dotenv>=1.2.1",
    "gradio>=5.0.0",
]

[project.optional-dependencies]
parse = [
    "modal>=0.73.0",
]

[tool.uv]
dev-dependencies = [
    "modal>=0.73.0",
]
